+++ vrpn_Spaceball.C
@@ -92,7 +92,7 @@
 // This routine writes out the characters slowly, so as not to
 // overburden the Spaceball (came from VRPN Magellan driver code)
 static	int	vrpn_write_slowly(int fd, unsigned char *buffer, size_t len, int MsecWait)
-{	int	i;
+{	size_t	i;
 
 	for (i = 0; i < len; i++) {
 		vrpn_SleepMsecs(MsecWait);
